new qasm code organize + tests
